requirements.txt
python-dotenv
requests
flask
openai
